+++ CHANGES.rst
@@ -24,6 +24,7 @@ Development
 - Improve rshell access over IP
 - Rename environment variable ``MCU_SERIAL_PORT`` to ``MCU_PORT``
   for configuring both USB/UART port and IP address
+- Use offset values for DS18B20 sensors from settings
 
 
 2019-06-22 0.5.1